src/Assistant/requirements.txt
langgraph==0.4.8
langchain[openai]==0.3.25
langchain-mcp-adapters==0.1.7
streamlit==1.45.1
python-dotenv==1.1.0
